refactor: auto-split test_locomotion.py
