+++ tests/unit/robotics/locomotion/test_gait_types.py
@@ -1,0 +1,99 @@
+"""Unit tests for locomotion module.
+
+Tests cover:
+    - Gait types and parameters
+    - ZMP computation
+    - Gait state machine
+    - Footstep planning
+"""
+
+from __future__ import annotations
+
+import numpy as np
+import pytest
+from numpy.testing import assert_allclose, assert_array_equal
+from src.robotics.locomotion.footstep_planner import (
+    Footstep,
+    FootstepPlan,
+    FootstepPlanner,
+)
+from src.robotics.locomotion.gait_state_machine import (
+    GaitEvent,
+    GaitState,
+    GaitStateMachine,
+)
+from src.robotics.locomotion.gait_types import (
+    GaitParameters,
+    GaitPhase,
+    GaitType,
+    SupportState,
+    create_run_parameters,
+    create_stand_parameters,
+    create_walk_parameters,
+)
+from src.robotics.locomotion.zmp_computer import (
+    ZMPComputer,
+)
+
+
+class TestGaitTypes:
+    """Tests for gait type enumerations."""
+
+    def test_gait_type_values(self) -> None:
+        """Test GaitType enum values exist."""
+        assert GaitType.STAND is not None
+        assert GaitType.WALK is not None
+        assert GaitType.RUN is not None
+        assert GaitType.TROT is not None
+
+    def test_gait_phase_values(self) -> None:
+        """Test GaitPhase enum values exist."""
+        assert GaitPhase.DOUBLE_SUPPORT is not None
+        assert GaitPhase.LEFT_SUPPORT is not None
+        assert GaitPhase.RIGHT_SUPPORT is not None
+        assert GaitPhase.FLIGHT is not None
+
+    def test_support_state_values(self) -> None:
+        """Test SupportState enum values."""
+        assert SupportState.DOUBLE_SUPPORT_CENTERED is not None
+        assert SupportState.SINGLE_SUPPORT_LEFT is not None
+        assert SupportState.SINGLE_SUPPORT_RIGHT is not None
+
+
+class MockHumanoidEngine:
+    """Mock humanoid engine for ZMP tests."""
+
+    def __init__(self) -> None:
+        self._com = np.array([0.0, 0.0, 0.9])
+        self._com_vel = np.zeros(3)
+        self._mass = 70.0
+
+    def get_state(self) -> tuple[np.ndarray, np.ndarray]:
+        return np.zeros(10), np.zeros(10)
+
+    def set_state(self, q: np.ndarray, v: np.ndarray) -> None:
+        pass
+
+    def compute_mass_matrix(self) -> np.ndarray:
+        return np.eye(10)
+
+    def compute_bias_forces(self) -> np.ndarray:
+        return np.zeros(10)
+
+    def compute_gravity_forces(self) -> np.ndarray:
+        return np.zeros(10)
+
+    def compute_jacobian(self, body_name: str) -> dict | None:
+        return {"linear": np.zeros((3, 10)), "angular": np.zeros((3, 10))}
+
+    def get_time(self) -> float:
+        return 0.0
+
+    def get_com_position(self) -> np.ndarray:
+        return self._com.copy()
+
+    def get_com_velocity(self) -> np.ndarray:
+        return self._com_vel.copy()
+
+    def get_total_mass(self) -> float:
+        return self._mass
